+++ configure.py
@@ -1326,6 +1326,7 @@ scylla_core = (['message/messaging_service.cc',
                 'ent/ldap/ldap_connection.cc',
                 'reader_concurrency_semaphore.cc',
                 'sstables_loader.cc',
+                'sstables_loader_helpers.cc',
                 'utils/utf8.cc',
                 'utils/ascii.cc',
                 'utils/like_matcher.cc',
